+++ packages/networks/src/networks/optimism.ts
@@ -54,6 +54,11 @@ export const optimism: NetworkConfig = {
         id: HookType.PROMO_CODE_CAPPED,
         name: 'Discount code with caps',
       },
+      {
+        address: '0x7217b772788374391e890b773e6b8B7101b5Acde',
+        id: HookType.PASSWORD_CAPPED,
+        name: 'Passwords with caps. Multiple passwords can be used per contract',
+      },
     ],
   },
   id: 10,
@@ -85,7 +90,7 @@ export const optimism: NetworkConfig = {
       'https://api.studio.thegraph.com/query/65299/unlock-protocol-optimism/version/latest',
     studioName: 'unlock-protocol-optimism',
   },
-  swapPurchaser: '0x72381052e4F7765A00a403891420BF75876c75bB',
+  swapPurchaser: '0x1bd356194d97297F77e081fFFAB97b57297E93e4',
   tokens: [
     {
       address: '0x7F5c764cBc14f9669B88837ca1490cCa17c31607',