+++ packages/unlock-js/src/PublicLock/v10/purchaseKeys.js
@@ -94,6 +94,7 @@ export default async function (options, transactionOptions = {}, callback) {
         ? unlockSwapPurchaserContract?.estimateGas?.swapAndCall(
             lockAddress,
             swap.srcTokenAddress || ZERO,
+            totalPrice,
             swap.amountInMax,
             swap.uniswapRouter,
             swap.swapCallData,
@@ -128,6 +129,7 @@ export default async function (options, transactionOptions = {}, callback) {
     ? unlockSwapPurchaserContract?.populateTransaction?.swapAndCall(
         lockAddress,
         swap.srcTokenAddress || ZERO,
+        totalPrice,
         swap.amountInMax,
         swap.uniswapRouter,
         swap.swapCallData,