+++ unlock-app/src/components/content/certification/CertificationForm.tsx
@@ -19,11 +19,13 @@ import { SelectCurrencyModal } from '~/components/interface/locks/Create/modals/
 import { CryptoIcon } from '@unlock-protocol/crypto-icon'
 import { useImageUpload } from '~/hooks/useImageUpload'
 import { BalanceWarning } from '~/components/interface/locks/Create/elements/BalanceWarning'
+import { NetworkWarning } from '~/components/interface/locks/Create/elements/NetworkWarning'
 import { getAccountTokenBalance } from '~/hooks/useAccount'
 import { Web3Service } from '@unlock-protocol/unlock-js'
 import { useQuery } from '@tanstack/react-query'
 import { useRouter } from 'next/router'
 import { useAvailableNetworks } from '~/utils/networks'
+import Link from 'next/link'
 
 // TODO replace with zod, but only once we have replaced Lock and MetadataFormData as well
 export interface NewCertificationForm {
@@ -115,11 +117,21 @@ export const CertificationForm = ({ onSubmit }: FormProps) => {
 
   const NetworkDescription = () => {
     return (
-      <p>
-        This is the network on which your certification contract will be
-        deployed.{' '}
-        {details.network && <>{networkDescription(details.network)}</>}
-      </p>
+      <div className="flex flex-col gap-2">
+        <p>{details.network && <>{networkDescription(details.network)}</>}</p>
+        <p>
+          This is the network on which your certification contract will be
+          deployed.{' '}
+          <Link
+            className="underline text-brand-ui-primary "
+            target="_blank"
+            href="https://unlock-protocol.com/guides/how-to-choose-a-network-for-your-smart-contract-deployment/"
+          >
+            Read our guide
+          </Link>{' '}
+          on how to chose the right network.
+        </p>
+      </div>
     )
   }
 
@@ -223,7 +235,7 @@ export const CertificationForm = ({ onSubmit }: FormProps) => {
                 <div>
                   <div className="flex items-center justify-between">
                     <label className="px-1 mb-2 text-base" htmlFor="">
-                      Certification duration (in days)
+                      Certification duration (days)
                     </label>
                     <ToggleSwitch
                       title="Forever"
@@ -269,14 +281,15 @@ export const CertificationForm = ({ onSubmit }: FormProps) => {
                 defaultValue={network}
                 description={<NetworkDescription />}
               />
-              <div className="mb-4">
-                {noBalance && (
+              <NetworkWarning network={details.network} />
+              {noBalance && (
+                <div className="mb-4">
                   <BalanceWarning
                     network={details.network!}
                     balance={parseFloat(balance)}
                   />
-                )}
-              </div>
+                </div>
+              )}
             </div>
           </Disclosure>