+++ unlock-app/src/components/interface/checkout/main/Promo.tsx
@@ -14,9 +14,14 @@ import { getEthersWalletFromPassword } from '~/utils/strings'
 import { Web3Service } from '@unlock-protocol/unlock-js'
 import { useDebounce } from 'react-use'
 import LoadingIcon from '../../Loading'
+import { useRouter } from 'next/router'
 interface Props {
   injectedProvider: unknown
   checkoutService: CheckoutService
+  recipients: string[]
+  lock: any
+  promoCode?: string
+  send: (obj: any) => void
 }
 
 interface FormData {
@@ -25,13 +30,33 @@ interface FormData {
 
 const web3Service = new Web3Service(networks)
 
-export function Promo({ injectedProvider, checkoutService }: Props) {
+export const computePromoData = async (promo: string, recipients: string[]) => {
+  const privateKeyAccount = await getEthersWalletFromPassword(promo)
+  return Promise.all(
+    recipients.map((address) => {
+      const messageHash = ethers.utils.solidityKeccak256(
+        ['string'],
+        [address.toLowerCase()]
+      )
+      const messageHashBinary = ethers.utils.arrayify(messageHash)
+      return privateKeyAccount.signMessage(messageHashBinary)
+    })
+  )
+}
+
+export function PromoContent({
+  recipients,
+  lock,
+  promoCode,
+  send,
+  injectedProvider,
+  checkoutService,
+}: Props) {
   const { account } = useAuth()
-  const [state, send] = useActor(checkoutService)
   const [hookAddress, setHookAddress] = useState<string>()
-  const [code, setCode] = useState<string>()
+  const [code, setCode] = useState<string | undefined>(promoCode)
+  const [promoCodeLoading, setPromoCodeLoading] = useState<boolean>(false)
   const [promoCodeDetails, setPromoCodeDetails] = useState<any>()
-  const { recipients, lock } = state.context
   const {
     register,
     handleSubmit,
@@ -54,6 +79,7 @@ export function Promo({ injectedProvider, checkoutService }: Props) {
   useDebounce(
     async () => {
       if (hookAddress && code) {
+        setPromoCodeLoading(true)
         const privateKeyFromAccount = await getEthersWalletFromPassword(code)
         const promoCodeDetails = await web3Service.getDiscountHookWithCapValues(
           {
@@ -64,26 +90,17 @@ export function Promo({ injectedProvider, checkoutService }: Props) {
           }
         )
         setPromoCodeDetails(promoCodeDetails)
+        setPromoCodeLoading(false)
       }
     },
-    300,
-    [code]
+    100,
+    [code, hookAddress, lock]
   )
 
   const onSubmit = async (formData: FormData) => {
     try {
       const { promo } = formData
-      const privateKeyAccount = await getEthersWalletFromPassword(promo)
-      const data = await Promise.all(
-        users.map((address) => {
-          const messageHash = ethers.utils.solidityKeccak256(
-            ['string'],
-            [address.toLowerCase()]
-          )
-          const messageHashBinary = ethers.utils.arrayify(messageHash)
-          return privateKeyAccount.signMessage(messageHashBinary)
-        })
-      )
+      const data = await computePromoData(promo, users)
       send({
         type: 'SUBMIT_DATA',
         data,
@@ -102,22 +119,24 @@ export function Promo({ injectedProvider, checkoutService }: Props) {
 
   const iconRight = isLoading
     ? LoadingIcon
-    : () => {
-        if (hasDiscount) {
-          return (
-            <Badge variant="green" size="tiny">
-              {promoCodeDetails.discount / 100}% Discount
-            </Badge>
-          )
-        } else if (promoCodeDetails?.discount > 0) {
-          return (
-            <Badge variant="dark" size="tiny">
-              Code expired
-            </Badge>
-          )
-        }
-        return null
-      }
+    : hasDiscount
+    ? () => (
+        <Badge variant="green" size="tiny">
+          {promoCodeDetails.discount / 100}% Discount
+        </Badge>
+      )
+    : promoCodeDetails?.discount > 0
+    ? () => (
+        <Badge variant="dark" size="tiny">
+          Code expired
+        </Badge>
+      )
+    : undefined
+
+  let error = errors.promo?.message
+  if (!error && code && promoCodeDetails?.discount === 0) {
+    error = "We couldn't find a discount for this code. "
+  }
 
   return (
     <Fragment>
@@ -128,11 +147,17 @@ export function Promo({ injectedProvider, checkoutService }: Props) {
             // @ts-ignore
             iconRight={iconRight}
             label="Enter Promo Code"
-            description="If you have a promo code to receive discounts, please enter it now."
+            description={
+              !promoCodeDetails?.discount
+                ? 'If you have a promo code to receive discounts, please enter it now.'
+                : ''
+            }
             type="text"
             size="small"
-            {...register('promo')}
-            error={errors.promo?.message}
+            {...register('promo', {
+              value: code,
+            })}
+            error={error}
             onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
               setCode(event.target.value)
               setPromoCodeDetails(undefined)
@@ -149,7 +174,7 @@ export function Promo({ injectedProvider, checkoutService }: Props) {
             type="submit"
             form="promo"
             className="w-full"
-            disabled={isSubmitting}
+            disabled={isSubmitting || promoCodeLoading}
             loading={isSubmitting}
             onClick={handleSubmit(onSubmit)}
           >
@@ -159,5 +184,34 @@ export function Promo({ injectedProvider, checkoutService }: Props) {
         <PoweredByUnlock />
       </footer>
     </Fragment>
+  )
+}
+
+interface PromoProps {
+  injectedProvider: unknown
+  checkoutService: CheckoutService
+}
+
+export function Promo({ injectedProvider, checkoutService }: PromoProps) {
+  const [state, send] = useActor(checkoutService)
+  const { query } = useRouter()
+  const { recipients, lock, paywallConfig } = state.context
+
+  let promoCode = ''
+  if (query?.promo) {
+    promoCode = query?.promo.toString()
+  } else if (typeof paywallConfig.locks[lock!.address].promo === 'string') {
+    promoCode = paywallConfig.locks[lock!.address].promo as string
+  }
+
+  return (
+    <PromoContent
+      recipients={recipients}
+      lock={lock}
+      promoCode={promoCode}
+      send={send}
+      injectedProvider={injectedProvider}
+      checkoutService={checkoutService}
+    />
   )
 }