+++ unlock-app/src/components/interface/checkout/main/checkoutMachine.ts
@@ -149,10 +149,6 @@ type Payment =
       route?: any
     }
   | {
-      method: 'universal_card'
-      cardId?: string
-    }
-  | {
       method: 'crossmint'
     }