+++ unlock-app/src/components/interface/locks/Settings/forms/UpdateHooksForm.tsx
@@ -7,12 +7,14 @@ import { useAuth } from '~/contexts/AuthenticationContext'
 import { Hook, HookName, HookType } from '@unlock-protocol/types'
 import { CustomContractHook } from './hooksComponents/CustomContractHook'
 import { PasswordContractHook } from './hooksComponents/PasswordContractHook'
+import { PasswordCappedContractHook } from './hooksComponents/PasswordCappedContractHook'
 import { DEFAULT_USER_ACCOUNT_ADDRESS } from '~/constants'
 import { useConfig } from '~/utils/withConfig'
 import { CaptchaContractHook } from './hooksComponents/CaptchaContractHook'
 import { GuildContractHook } from './hooksComponents/GuildContractHook'
 import { PromoCodeHook } from './hooksComponents/PromoCodeHook'
 import { useCustomHook } from '~/hooks/useCustomHooks'
+import { GitcoinContractHook } from './hooksComponents/GitcoinContractHook'
 
 interface UpdateHooksFormProps {
   lockAddress: string
@@ -78,6 +80,11 @@ export const HookMapping: Record<FormPropsKey, HookValueProps> = {
         component: (args) => <PasswordContractHook {...args} />,
       },
       {
+        label: 'Password with caps',
+        value: HookType.PASSWORD_CAPPED,
+        component: (args) => <PasswordCappedContractHook {...args} />,
+      },
+      {
         label: 'Captcha required',
         value: HookType.CAPTCHA,
         component: (args) => <CaptchaContractHook {...args} />,
@@ -91,6 +98,11 @@ export const HookMapping: Record<FormPropsKey, HookValueProps> = {
         label: 'Discount code',
         value: HookType.PROMO_CODE_CAPPED,
         component: (args) => <PromoCodeHook {...args} />,
+      },
+      {
+        label: 'Gitcoin Passport verification',
+        value: HookType.GITCOIN,
+        component: (args) => <GitcoinContractHook {...args} />,
       },
     ],
   },